feat: rename setup to registration to make a place for true setup
The setup concept is now the initial single-person creation
of the game and selection of categories.
