+++ src/components/registration/contestant-registration.tsx
