+++ src/components/registration/current-registration.tsx
