+++ src/components/registration/display-registration.tsx
