+++ src/components/registration/host-registration.tsx
