+++ src/components/registration/index.ts
@@ -1,0 +1,5 @@
+export * from '../member-select';
+export * from './contestant-registration';
+export * from './current-registration';
+export * from './display-registration';
+export * from './host-registration';
