+++ src/components/setup/index.ts
@@ -1,7 +1,2 @@
 export * from './category-selector';
-export * from './contestant-registration';
-export * from './current-registration';
-export * from './display-registration';
-export * from './host-registration';
-export * from './member-select';
 export * from './view-category';
