+++ src/hooks/__tests__/use-game-phase-registration-state.test.tsx
@@ -1,6 +1,6 @@
 import { act, renderHook } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
-import useGamePhaseSetupState from '@/hooks/use-game-phase-setup-state';
+import useGamePhaseRegistrationState from '@/hooks/use-game-phase-registration-state';
 
 const GAME_ID = 'test-game';
 const SESSION_A = 'session-a';
@@ -8,37 +8,37 @@ const SESSION_B = 'session-b';
 
 describe('registerAsDisplay', () => {
 	it('registers the current session as display', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsDisplay());
 		expect(result.current.display).toBe(SESSION_A);
 	});
 
 	it('sets hasDisplay to true', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsDisplay());
 		expect(result.current.hasDisplay).toBe(true);
 	});
 
 	it('throws when already registered as display', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsDisplay());
 		expect(() => act(() => result.current.registerAsDisplay())).toThrow();
 	});
 
 	it('throws when already registered as host', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsHost());
 		expect(() => act(() => result.current.registerAsDisplay())).toThrow();
 	});
 
 	it('throws when already registered as contestant', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsContestant('Alice', undefined));
 		expect(() => act(() => result.current.registerAsDisplay())).toThrow();
 	});
 
 	it('throws when another session is already the display', () => {
-		const { result, rerender } = renderHook(({ sessionId }) => useGamePhaseSetupState(sessionId, GAME_ID), {
+		const { result, rerender } = renderHook(({ sessionId }) => useGamePhaseRegistrationState(sessionId, GAME_ID), {
 			initialProps: { sessionId: SESSION_A },
 		});
 		act(() => result.current.registerAsDisplay());
@@ -49,33 +49,33 @@ describe('registerAsDisplay', () => {
 
 describe('unregisterAsDisplay', () => {
 	it('clears display', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsDisplay());
 		act(() => result.current.unregisterAsDisplay());
 		expect(result.current.display).toBeUndefined();
 	});
 
 	it('sets hasDisplay to false', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsDisplay());
 		act(() => result.current.unregisterAsDisplay());
 		expect(result.current.hasDisplay).toBe(false);
 	});
 
 	it('silently succeeds when not registered as anything', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		expect(() => act(() => result.current.unregisterAsDisplay())).not.toThrow();
 	});
 
 	it('silently succeeds and leaves host intact when registered as host', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsHost());
 		act(() => result.current.unregisterAsDisplay());
 		expect(result.current.host).toBe(SESSION_A);
 	});
 
 	it('silently succeeds and leaves contestants intact when registered as contestant', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsContestant('Alice', undefined));
 		act(() => result.current.unregisterAsDisplay());
 		expect(result.current.contestants).toHaveLength(1);
@@ -84,37 +84,37 @@ describe('unregisterAsDisplay', () => {
 
 describe('registerAsHost', () => {
 	it('registers the current session as host', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsHost());
 		expect(result.current.host).toBe(SESSION_A);
 	});
 
 	it('sets hasHost to true', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsHost());
 		expect(result.current.hasHost).toBe(true);
 	});
 
 	it('throws when already registered as host', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsHost());
 		expect(() => act(() => result.current.registerAsHost())).toThrow();
 	});
 
 	it('throws when already registered as display', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsDisplay());
 		expect(() => act(() => result.current.registerAsHost())).toThrow();
 	});
 
 	it('throws when already registered as contestant', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsContestant('Alice', undefined));
 		expect(() => act(() => result.current.registerAsHost())).toThrow();
 	});
 
 	it('throws when another session is already the host', () => {
-		const { result, rerender } = renderHook(({ sessionId }) => useGamePhaseSetupState(sessionId, GAME_ID), {
+		const { result, rerender } = renderHook(({ sessionId }) => useGamePhaseRegistrationState(sessionId, GAME_ID), {
 			initialProps: { sessionId: SESSION_A },
 		});
 		act(() => result.current.registerAsHost());
@@ -125,33 +125,33 @@ describe('registerAsHost', () => {
 
 describe('unregisterAsHost', () => {
 	it('clears host', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsHost());
 		act(() => result.current.unregisterAsHost());
 		expect(result.current.host).toBeUndefined();
 	});
 
 	it('sets hasHost to false', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsHost());
 		act(() => result.current.unregisterAsHost());
 		expect(result.current.hasHost).toBe(false);
 	});
 
 	it('silently succeeds when not registered as anything', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		expect(() => act(() => result.current.unregisterAsHost())).not.toThrow();
 	});
 
 	it('silently succeeds and leaves display intact when registered as display', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsDisplay());
 		act(() => result.current.unregisterAsHost());
 		expect(result.current.display).toBe(SESSION_A);
 	});
 
 	it('silently succeeds and leaves contestants intact when registered as contestant', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsContestant('Alice', undefined));
 		act(() => result.current.unregisterAsHost());
 		expect(result.current.contestants).toHaveLength(1);
@@ -160,31 +160,31 @@ describe('unregisterAsHost', () => {
 
 describe('registerAsContestant', () => {
 	it('adds the current session to contestants', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsContestant('Alice', undefined));
 		expect(result.current.contestants).toContainEqual({ sessionId: SESSION_A, name: 'Alice', userId: undefined });
 	});
 
 	it('stores the provided userId', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsContestant('Alice', 'user-1'));
 		expect(result.current.contestants[0].userId).toBe('user-1');
 	});
 
 	it('throws when already registered as contestant', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsContestant('Alice', undefined));
 		expect(() => act(() => result.current.registerAsContestant('Alice', undefined))).toThrow();
 	});
 
 	it('throws when already registered as host', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsHost());
 		expect(() => act(() => result.current.registerAsContestant('Alice', undefined))).toThrow();
 	});
 
 	it('throws when already registered as display', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsDisplay());
 		expect(() => act(() => result.current.registerAsContestant('Alice', undefined))).toThrow();
 	});
@@ -192,26 +192,26 @@ describe('registerAsContestant', () => {
 
 describe('unregisterAsContestant', () => {
 	it('removes the current session from contestants', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsContestant('Alice', undefined));
 		act(() => result.current.unregisterAsContestant());
 		expect(result.current.contestants).toHaveLength(0);
 	});
 
 	it('silently succeeds when not registered as anything', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		expect(() => act(() => result.current.unregisterAsContestant())).not.toThrow();
 	});
 
 	it('silently succeeds and leaves display intact when registered as display', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsDisplay());
 		act(() => result.current.unregisterAsContestant());
 		expect(result.current.display).toBe(SESSION_A);
 	});
 
 	it('silently succeeds and leaves host intact when registered as host', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsHost());
 		act(() => result.current.unregisterAsContestant());
 		expect(result.current.host).toBe(SESSION_A);
@@ -220,44 +220,44 @@ describe('unregisterAsContestant', () => {
 
 describe('currentUserRole', () => {
 	it('is undefined before any registration', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		expect(result.current.currentUserRole).toBeUndefined();
 	});
 
 	it('is display after registerAsDisplay', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsDisplay());
 		expect(result.current.currentUserRole).toBe('display');
 	});
 
 	it('is host after registerAsHost', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsHost());
 		expect(result.current.currentUserRole).toBe('host');
 	});
 
 	it('is contestant after registerAsContestant', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsContestant('Alice', undefined));
 		expect(result.current.currentUserRole).toBe('contestant');
 	});
 
 	it('is undefined after unregistering as display', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsDisplay());
 		act(() => result.current.unregisterAsDisplay());
 		expect(result.current.currentUserRole).toBeUndefined();
 	});
 
 	it('is undefined after unregistering as host', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsHost());
 		act(() => result.current.unregisterAsHost());
 		expect(result.current.currentUserRole).toBeUndefined();
 	});
 
 	it('is undefined after unregistering as contestant', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		act(() => result.current.registerAsContestant('Alice', undefined));
 		act(() => result.current.unregisterAsContestant());
 		expect(result.current.currentUserRole).toBeUndefined();
@@ -266,19 +266,19 @@ describe('currentUserRole', () => {
 
 describe('initial state', () => {
 	it('has no display', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		expect(result.current.display).toBeUndefined();
 		expect(result.current.hasDisplay).toBe(false);
 	});
 
 	it('has no host', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		expect(result.current.host).toBeUndefined();
 		expect(result.current.hasHost).toBe(false);
 	});
 
 	it('has no contestants', () => {
-		const { result } = renderHook(() => useGamePhaseSetupState(SESSION_A, GAME_ID));
+		const { result } = renderHook(() => useGamePhaseRegistrationState(SESSION_A, GAME_ID));
 		expect(result.current.contestants).toHaveLength(0);
 	});
 });
