+++ src/models/enums.ts
@@ -1,2 +1,2 @@
-export const gamePhaseEnum = ['SETUP', 'PLAYING', 'FINISHED'] as const;
+export const gamePhaseEnum = ['SETUP', 'REGISTRATION', 'PLAYING', 'FINISHED'] as const;
 export const gameStageEnum = ['SINGLE', 'DOUBLE', 'TRIPLE', 'FINAL'] as const;
