+++ src/pages/games/routes.tsx
@@ -5,12 +5,12 @@ import Pages__Games__Edit from './edit';
 import Pages__Games__Listing from './listing';
 import Pages__Games__New from './new';
 import Pages__Games__Play from './play';
-import Pages__Games__Setup from './setup';
+import Pages__Games__Registration from './registration';
 
 export default [
 	route('/listing', [requireAuthentication, requirePermissions('games:read'), Pages__Games__Listing]),
 	route('/new', [requireAuthentication, requirePermissions('games:create'), Pages__Games__New]),
 	route('/:gameId/play', [ensureSession, requirePermissions('games:read'), Pages__Games__Play]),
 	route('/:gameId/edit', [requireAuthentication, requirePermissions('games:update'), Pages__Games__Edit]),
-	route('/:gameId/setup', [ensureSession, requirePermissions('games:read'), Pages__Games__Setup]),
+	route('/:gameId/registration', [ensureSession, requirePermissions('games:read'), Pages__Games__Registration]),
 ];
