+++ src/styles/main.css
@@ -7,6 +7,7 @@
 @import "./view-contestant.less";
 @import "./view-display.less";
 @import "./view-host.less";
+@import "./view-registration.less";
 @import "./view-setup.less";
 @import "./welcome.less";
