+++ src/styles/view-registration.less
@@ -1,0 +1,11 @@
+.view-game-registration {
+	display: flex;
+	.view-game-registration-actions,
+	.view-game-registration-qr-code {
+		flex: 1 0 50%;
+	}
+	.view-game-registration-qr-code {
+		padding: 1rem;
+		text-align: center;
+	}
+}
